Refactor role-related functions and add role name utility function
This commit refactors the role-related functions in the `Section.jsx` and `Card.jsx` files to a separate `utils.js` file. The `getUsersStats` function is moved to the `utils.js` file and exported. Additionally, a new utility function `getRoleName` is added to the `utils.js` file, which returns the role name based on the number of roles a user has.

This refactoring improves code organization and reusability by separating the role-related logic into a separate file.